feat: added menu card and done menu page
